refactor: clean up unused code
