+++ api/package.json
@@ -9,7 +9,7 @@
   "private": true,
   "scripts": {
     "start": "node dist/index.js",
-    "build": "npm run tsc && swc src -d dist",
+    "build": "run-p tsc swc",
     "generate": "run-p generate:*",
     "generate:prisma": "prisma generate",
     "generate:gql": "graphql-codegen --config codegen.config.ts",
@@ -29,7 +29,8 @@
     "test:integration": "npm run integration-test --",
     "posttest:integration": "npm run docker:down",
     "test:integration:headless": "npm run integration-test --",
-    "tsc": "tsc"
+    "tsc": "tsc",
+    "swc": "swc src -d dist"
   },
   "optionalDependencies": {
     "@swc/core-linux-x64-musl": "^1.3.11"
