+++ api/prisma/schema.prisma
@@ -17,24 +17,23 @@ model GoogleDocument {
 }
 
 model User {
-  id                      String       @id @default(uuid())
-  feideId                 String       @unique
-  createdAt               DateTime     @default(now())
-  updatedAt               DateTime     @updatedAt
-  lastLogin               DateTime     @default(now())
-  firstLogin              Boolean      @default(true)
+  id                      String    @id @default(uuid())
+  feideId                 String    @unique
+  createdAt               DateTime  @default(now())
+  updatedAt               DateTime  @updatedAt
+  lastLogin               DateTime  @default(now())
+  firstLogin              Boolean   @default(true)
   firstName               String
   lastName                String
-  email                   String       @unique
-  username                String       @unique
+  email                   String    @unique
+  username                String    @unique
   roles                   Role[]
-  permissions             Permission[]
   Event                   Event[]
   graduationYear          Int?
   graduationYearUpdatedAt DateTime?
   allergies               String?
   phoneNumber             String?
-  isSuperUser             Boolean      @default(false)
+  isSuperUser             Boolean   @default(false)
 }
 
 model Organization {
@@ -58,21 +57,11 @@ model Event {
 }
 
 model Role {
-  id          String       @id @default(uuid())
-  createdAt   DateTime     @default(now())
-  name        String       @unique
-  permissions Permission[]
-  User        User?        @relation(fields: [userId], references: [id])
-  userId      String?
-}
-
-model Permission {
-  id     String  @id @default(uuid())
-  name   String  @unique
-  role   Role?   @relation(fields: [roleId], references: [id])
-  roleId String?
-  user   User?   @relation(fields: [userId], references: [id])
-  userId String?
+  id        String   @id @default(uuid())
+  createdAt DateTime @default(now())
+  name      String   @unique
+  User      User?    @relation(fields: [userId], references: [id])
+  userId    String?
 }
 
 // Cabins
