+++ api/src/container.ts
@@ -8,15 +8,19 @@ import { Context } from "@/graphql";
 import * as Repositories from "@/repositories";
 import * as Services from "@/services";
 
-// register Prisma
-container.register<PrismaClient>(CoreTypes.Prisma, { useValue: Prisma });
+function initializeContainer() {
+  // register Prisma
+  container.register<PrismaClient>(CoreTypes.Prisma, { useValue: Prisma });
 
-// register Postmark
-container.register<IMailClient>(CoreTypes.MailClient, { useValue: Postmark });
+  // register Postmark
+  container.register<IMailClient>(CoreTypes.MailClient, { useValue: Postmark });
 
-// register Context
-Context.register(container);
-// register Repositories
-Repositories.register(container);
-// register Services
-Services.register(container);
+  // register Context
+  Context.register(container);
+  // register Repositories
+  Repositories.register(container);
+  // register Services
+  Services.register(container);
+}
+
+export { initializeContainer };
