+++ api/src/graphql/context.ts
@@ -1,20 +1,18 @@
-import { ExpressContextFunctionArgument } from "@apollo/server/dist/esm/express4";
+import { ExpressContextFunctionArgument } from "@apollo/server/express4";
 import { DependencyContainer, inject, injectable } from "tsyringe";
 
-import { IAuthService, ICabinService, IPermissionService, IUserService, Types } from "@/services";
+import { IAuthService, ICabinService, IUserService, Types } from "@/services";
 
 export const Type = Symbol.for("ContextProvider");
 
 export interface IContextProvider {
   userService: IUserService;
-  permissionService: IPermissionService;
   cabinService: ICabinService;
   authService: IAuthService;
 }
 
 export interface IContext extends ExpressContextFunctionArgument {
   userService: IUserService;
-  permissionService: IPermissionService;
   cabinService: ICabinService;
   authService: IAuthService;
 }
@@ -23,8 +21,6 @@ export interface IContext extends ExpressContextFunctionArgument {
 class ContextProvider implements IContextProvider {
   public constructor(
     @inject(Types.UserService) public userService: IUserService,
-    @inject(Types.PermissionService)
-    public permissionService: IPermissionService,
     @inject(Types.CabinService) public cabinService: ICabinService,
     @inject(Types.AuthService) public authService: IAuthService
   ) {}
