+++ api/src/graphql/index.ts
@@ -1,3 +1,3 @@
-export { default as Context } from "@/graphql/context";
-export { default as resolvers } from "@/graphql/resolvers";
-export { default as typeDefs } from "@/graphql/type-defs";
+export { default as Context } from "./context";
+export { default as resolvers } from "./resolvers";
+export { default as typeDefs } from "./type-defs";
