--- api/src/graphql/permissions/index.ts
@@ -1,2 +1,0 @@
-export { default as typeDefs } from "./type-defs";
-export { default as resolvers } from "./resolvers";
