--- api/src/graphql/permissions/type-defs.ts
@@ -1,14 +1,0 @@
-import gql from "graphql-tag";
-
-const typeDefs = gql`
-  type Query {
-    hasPermission(permission: String!): Boolean!
-  }
-
-  type Permission {
-    id: ID!
-    name: String!
-  }
-`;
-
-export default typeDefs;
