+++ api/src/graphql/resolvers.ts
@@ -1,18 +1,11 @@
 import { mergeResolvers } from "@graphql-tools/merge";
 
-import { resolvers as authResolvers } from "@/graphql/auth";
-import { resolvers as cabinResolvers } from "@/graphql/cabins";
-import { Resolvers } from "@/graphql/generated/types";
-import { resolvers as permissionResolvers } from "@/graphql/permissions";
-import { resolvers as scalarResolvers } from "@/graphql/scalars";
-import { resolvers as userResolvers } from "@/graphql/users";
+import { resolvers as authResolvers } from "./auth";
+import { resolvers as cabinResolvers } from "./cabins";
+import { Resolvers } from "./generated/types";
+import { resolvers as scalarResolvers } from "./scalars";
+import { resolvers as userResolvers } from "./users";
 
-const resolvers: Resolvers = mergeResolvers([
-  userResolvers,
-  cabinResolvers,
-  scalarResolvers,
-  authResolvers,
-  permissionResolvers,
-]);
+const resolvers: Resolvers = mergeResolvers([userResolvers, cabinResolvers, scalarResolvers, authResolvers]);
 
 export default resolvers;
