+++ api/src/graphql/scalars/index.ts
@@ -1,6 +1,6 @@
 import { DateTimeResolver } from "graphql-scalars";
 
-import { Resolvers } from "@/graphql/generated/types";
+import { Resolvers } from "../generated/types";
 
 export { default as typeDefs } from "./type-defs";
