+++ api/src/graphql/type-defs.ts
@@ -1,11 +1,10 @@
 import { mergeTypeDefs } from "@graphql-tools/merge";
 
-import { typeDefs as authTypes } from "@/graphql/auth";
-import { typeDefs as cabinTypes } from "@/graphql/cabins";
-import { typeDefs as permissionTypes } from "@/graphql/permissions";
-import { typeDefs as scalarTypes } from "@/graphql/scalars";
-import { typeDefs as userTypes } from "@/graphql/users";
+import { typeDefs as authTypes } from "./auth";
+import { typeDefs as cabinTypes } from "./cabins";
+import { typeDefs as scalarTypes } from "./scalars";
+import { typeDefs as userTypes } from "./users";
 
-export const typeDefs = mergeTypeDefs([userTypes, permissionTypes, cabinTypes, scalarTypes, authTypes]);
+export const typeDefs = mergeTypeDefs([userTypes, cabinTypes, scalarTypes, authTypes]);
 
 export default typeDefs;
