+++ api/src/repositories/types.ts
@@ -1,5 +1,4 @@
 export const Types = {
   UserRepository: Symbol.for("UserRepository"),
-  PermissionRepository: Symbol.for("PermissionRepository"),
   CabinRepsitory: Symbol.for("CabinRepsitory"),
 };
