--- api/src/services/permissions/types.ts
@@ -1,7 +1,0 @@
-import { Prisma } from "@prisma/client";
-
-type Entity = Lowercase<Prisma.ModelName>;
-type Scope = "read" | "write";
-type Identifier = string;
-
-export type PermissionString = `${Scope}:${Entity}:${Identifier}` | `${Scope}:${Entity}`;
