+++ api/src/services/types.ts
@@ -1,6 +1,5 @@
 export const Types = {
   UserService: Symbol.for("UserService"),
-  PermissionService: Symbol.for("PermissionService"),
   MailService: Symbol.for("MailService"),
   CabinService: Symbol.for("CabinService"),
   AuthService: Symbol.for("AuthService"),
