+++ api/tsconfig.json
@@ -27,7 +27,5 @@
   },
   "include": ["src/**/*.ts"],
   "exclude": ["node_modules"],
-  "files": [
-    "src/index.ts"
-  ]
+  "files": ["src/index.ts"]
 }
